+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - pause() and unpause() in Dlogos
 - Error.sol
 - deployment and upgrade scripts
+- create a new [SplitsV2 PushSplit](https://docs.splits.org/core/split-v2) in Dlogos distribution function
 
 ## [2.0.1] - 2024-06-03
 ### Added